added dtos, mappers, services

+++ car_rental/src/main/java/com/team_spak/car_rental/repository/CarRepository.java
@@ -1,7 +1,7 @@
 package com.team_spak.car_rental.repository;
 
-import com.team_spak.car_rental.model.Car;
+import com.team_spak.car_rental.model.entity.Car;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarRepository  extends JpaRepository<Car, Integer> {
+public interface CarRepository  extends JpaRepository<Car, Long> {
 }
