+++ car_rental/src/main/java/com/team_spak/car_rental/repository/ReviewRepository.java
@@ -1,6 +1,6 @@
 package com.team_spak.car_rental.repository;
-import com.team_spak.car_rental.model.Review;
+import com.team_spak.car_rental.model.entity.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
+public interface ReviewRepository extends JpaRepository<Review, Long> {
 }
