+++ car_rental/src/main/java/com/team_spak/car_rental/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package com.team_spak.car_rental.repository;
 
-import com.team_spak.car_rental.model.User;
+import com.team_spak.car_rental.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository  extends JpaRepository <User, Integer> {
+public interface UserRepository  extends JpaRepository <User, Long> {
 }
